ui: responsive upload
